// src/typings/yapi/ImChatGroupAdministratorV1GetApi.d.ts
/* prettier-ignore-start */
/* tslint:disable */
/* eslint-disable */

/* 该文件由 yapi-to-typescript 自动生成，请勿直接修改！！！ */

/**
 * 接口 [群管理员列表↗](https://yapi.nbttfc365.com/project/82/interface/api/11499) 的 **请求类型**
 *
 * @分类 [群组模块↗](https://yapi.nbttfc365.com/project/82/interface/api/cat_812)
 * @请求头 `GET /v1/im/chat/group/administrator`
 * @更新时间 `2023-08-22 11:10:56`
 */
export interface YapiGetV1ImChatGroupAdministratorApiRequest {
  /**
   * 群组ID
   */
  groupId: string
}

/**
 * 接口 [群管理员列表↗](https://yapi.nbttfc365.com/project/82/interface/api/11499) 的 **返回类型**
 *
 * @分类 [群组模块↗](https://yapi.nbttfc365.com/project/82/interface/api/cat_812)
 * @请求头 `GET /v1/im/chat/group/administrator`
 * @更新时间 `2023-08-22 11:10:56`
 */
export interface YapiGetV1ImChatGroupAdministratorApiResponse {
  /**
   * 200，成功，其他失败
   */
  code: number
  /**
   * 描述信息
   */
  message: string
  data: YapiGetV1ImChatGroupAdministratorListData[]
}
export interface YapiGetV1ImChatGroupAdministratorListData {
  /**
   * 用户ID
   */
  uid: number
  /**
   * 用户昵称
   */
  nickName: string
  /**
   * 用户头像
   */
  avatarPath: string
}

// 以下为自动生成的 api 请求，需要使用的话请手动复制到相应模块的 api 请求层
// import request, { MarkcoinRequest } from '@/plugins/request'

// /**
// * [群管理员列表↗](https://yapi.nbttfc365.com/project/82/interface/api/11499)
// **/
// export const getV1ImChatGroupAdministratorApiRequest: MarkcoinRequest<
//   YapiGetV1ImChatGroupAdministratorApiRequest,
//   YapiGetV1ImChatGroupAdministratorApiResponse['data']
// > = params => {
//   return request({
//     path: "/v1/im/chat/group/administrator",
//     method: "GET",
//     params
//   })
// }

/* prettier-ignore-end */